pyproject.toml
[project]
name = "mcper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.2.4",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
]
